backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.2.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytesseract==0.3.10
python-docx==1.1.0
pytz==2025.1
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2025.1
Werkzeug==3.1.3
sentence-transformers==2.7.0
transformers==4.40.0
Levenshtein==0.27.1
fuzzywuzzy==0.18.0
python-levenshtein==0.27.1
rapidfuzz==3.13.0
bcrypt ==4.3.0
flask-bcrypt ==1.0.1
PyJWT==2.10.1
Flask-JWT-Extended == 4.7.1
pymongo == 4.12.0
dnspython==2.7.0
python-dotenv==1.1.0
